requirements.txt
numpy==2.3.2
pandas==2.3.1
python-dotenv==1.1.1
pytorch3d==0.7.5
scikit_learn==1.3.0
scipy==1.16.1
torch==2.0.1
wimusim.egg==info
